Add ability to define custom blog admin route.
Moved config to lib to be accessible by fast-render,
and removed conflict in them.
Moved routes definition into startup function so that
the config definiton is defined before it runs.
Add api.use underscore to both instead of just client.
Underscore is used on both, so a app that doesn't require
it on the server would not work as expected.