Increase specificity on event that prevents the editor form submit
If we're just trying to prevent the form from being submit, there's
no need to stop propagation and default on *all* clicks on these
buttons. It may cause unwanted behavior in the future. Also,
`BlogEditor.make` runs on every auto-save, which in turn calls this
event bind, which sometimes caused the buttons to do nothing when
clicked since `stopPropagation` was bound multiple times. Also,
since the event is delegated directly to the button, if for some reason
we removed the buttons and put them back, the bindings would be gone.

By delegating the event to the `form`'s click event and checking
that the event originated from a `.mediumInsert`, we only block
the event when necessary.